+++ apps/backend/src/modules/users/libs/validation-schemas/validation-schemas.ts
@@ -1,4 +1,5 @@
 export {
+	notificationAnswersValidationSchema,
 	userForgotPasswordValidationSchema,
 	userResetPasswordValidationSchema,
 	userSignInValidationSchema,