+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/enums/wheel-center-distance.enum.ts
@@ -1,0 +1,7 @@
+const WheelCenterDistance = {
+	DOT: 10.9,
+	SUBLABEL: 11.6,
+	VALUE_LABEL: 12.4,
+} as const;
+
+export { WheelCenterDistance };