+++ apps/frontend/src/libs/components/balance-wheel-chart/libs/types/graphics-position.type.ts
@@ -1,0 +1,3 @@
+type GraphicsPosition = { x: number; y: number };
+
+export { type GraphicsPosition };