+++ apps/frontend/src/libs/components/components.ts
@@ -1,6 +1,7 @@
 export { App } from "./app/app.js";
 export { BalanceWheelChart } from "./balance-wheel-chart/balance-wheel-chart.js";
 export { Button } from "./button/button.js";
+export { Checkbox } from "./checkbox/checkbox.js";
 export { Header } from "./header/header.js";
 export { Icon } from "./icon/icon.js";
 export { Input } from "./input/input.js";