+++ apps/frontend/src/modules/quiz/slices/actions.ts
@@ -1,7 +1,10 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
-import { type QuizQuestionDto } from "~/modules/quiz/quiz.js";
+import {
+	type QuizQuestionDto,
+	type QuizScoresGetAllResponseDto,
+} from "~/modules/quiz/quiz.js";
 
 import { name as sliceName } from "./quiz.slice.js";
 
@@ -15,4 +18,14 @@ const getAllQuestions = createAsyncThunk<
 	return await quizApi.getAllQuestions();
 });
 
-export { getAllQuestions };
+const getScores = createAsyncThunk<
+	QuizScoresGetAllResponseDto,
+	undefined,
+	AsyncThunkConfig
+>(`${sliceName}/score`, async (_, { extra }) => {
+	const { quizApi } = extra;
+
+	return await quizApi.getScores();
+});
+
+export { getAllQuestions, getScores };