+++ apps/mobile/src/libs/components/planet/planet.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 
 import { LinearGradient } from "~/libs/components/components";
-import { colorToGradientColors } from "~/libs/maps/maps";
+import { colorToGradientColors, directionToGradient } from "~/libs/maps/maps";
 import { type StyleProp, type ViewStyle } from "~/libs/types/types";
 
-import { directionToGradient, sizeToStyles } from "./libs/maps/maps";
+import { sizeToStyles } from "./libs/maps/maps";
 import { styles } from "./styles";
 
 type Properties = {