+++ apps/mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -1,10 +1,12 @@
 const RootScreenName = {
 	BOTTOM_TABS_NAVIGATOR: "BottomTabsNavigator",
 	CHAT: "Chat",
+	ONBOARDING: "Onboarding",
 	QUIZ_ENTRY: "Quiz Entry",
 	SIGN_IN: "Sign In",
 	SIGN_UP: "Sign Up",
 	WELCOME: "Welcome",
+	WHEEL_LOADING: "Wheel Loading",
 } as const;
 
 export { RootScreenName };