+++ packages/shared/src/modules/users/libs/types/notification-answers-payload-dto.type.ts
@@ -1,0 +1,9 @@
+import { type ValueOf } from "../../../../libs/types/types.js";
+import { type NotificationFrequency } from "../enums/enums.js";
+
+type NotificationAnswersPayloadDto = {
+	notificationFrequency: ValueOf<typeof NotificationFrequency>;
+	userTaskDays: number[];
+};
+
+export { type NotificationAnswersPayloadDto };